+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `awc_civic_cookie_control_config` filter, to allow default config to be added to or over-ridden
+
 ## [0.2.2] - 2021-04-12
 
 ### Added